// src/types/role.ts
/*
 * User roles configuration
 */
export enum Role {
  Developer = 'Console Dev',
  User = 'Console User',
  Akademik = 'old-bpa',
  BAA = 'old-baa',
  Kaprodi = 'old-kaprodi',
  KoordinatorMK = 'old-subject-coordinator',
  Dosen = 'old-dosen',
  Wadek1 = 'old-wakil-dekan-i',
  Warek = 'old-warek-1',
  LAA = 'old-admin-laak'
}